give more control over timing

+++ gtraceroute/core/application/services.py
@@ -51,7 +51,7 @@ class RouteHop:
         self,
         dispatcher: RequestDispatcher,
         reply_watcher: ICMPReplyWatcher,
-        timeout: int = 1,
+        timeout: float = 1,
     ):
         try:
             async with asyncio.timeout(timeout):
